requirements.txt
numpy==1.19.5
pandas==1.3.4
requests==2.26.0
scikit-learn==1.0.1
streamlit==1.20.0
streamlit-folium==0.11.1
python-dotenv==0.19.1